:green_heart: Fixes OS gh actions (#115)
* :green_heart: Fixes OS gh actions

* :wrench: Updates flak8 and pytest cfg files

This commit updates the pytest and flake8 configuration entries:
 - The flak8 configuration settings are now in the setup.cfg file.
 - The pytest configuration settings are now in the pyproject.toml file.
 - Pytest now ignores the test in the simzoo folder as they are already
 tested in the subpackage.

* :green_heart: Fixes gh-actions pytest exit code problem

This commit fixes the exit code problem that was present in the github
actions. The tee command set the exit code to 0 even when pytest found a
error. The pipefail options is used to solve this.

* :bug: Fixes a small bug in the gh-action

The last commit added a syntax error to the linux gh-action. This commit
fixes this error.

* :wrench: Updates dev requirements

This commit adds the tuning packages to the dev requirements. Otherwise
the coverage test fails.

* :wrench: Updates pytest configuration file

This commit updates the pytest config file to exclude all the folders
other than the bayesian_learning_control folder. Further some of the
gh-action pytest arguments were added to the pyproject.toml.