:rewind: Removes gpl algorithms from master branch
In the previous commit I merged changes from the gpl branch into the
master branch. This was done since the two branches converged to much
and I needed some changes from the gpl branch. I however do not think it
is a smart idea to add the new (unfinished) gpl algorithms to the main branch
I will continue development on branch gpl2.