fix: remove 'simzoo' submodule
This commit removes the `simzoo` submodule since it is now available as a stand-alone package called
[stable-gym](https://github.com/rickstaa/stable-gym). This was done since having the package in this
package required a lot of complicated code and made the package harder to debug.

BREAKING CHANGE: The `simzoo` package is no longer included and has to be installed
separately through the [stable-gym](https://github.com/rickstaa/stable-gym) repository.