feat: improve 'eval_robustness' utility (#313)
This commit improves the 'eval_robustness' utility so that users no
longer need to modify their environments to use it with this utility.
The disturbers are now implemented as [gymnasium
wrappers](https://gymnasium.farama.org/api/wrappers/). Because of this,
they can be used with any [gymnasium
environment](https://gymnasium.farama.org/). If you want to add a new
disturber, you only have to ensure that it is a Python class that
inherits from the
[gym.Wrapper](https://github.com/Farama-Foundation/Gymnasium/blob/ba348890af369dad6e6093665fd4e54d8d4b4824/gymnasium/core.py#L246-L249)
class.