refactor(buffers): remove unused 'rew_dim' argument
This commit removes the `rew_dim` argument in the `ReplayBuffer` and
`TrajectoryBuffer` since it is
uncommon for Gymansium environments to have multi-dimensional reward
spaces. This code was added
because of a prototype we did some time ago.

BREAKING CHANGE: the `ReplayBuffer` and `TrajectoryBuffer` classes don't
take a `rew_dim` argument anymore.